Renamed External Functions module, changed Closing Prompt, moved Toolbar to  External Functions module, etc.
