+++ .gitignore
@@ -4,3 +4,4 @@ save/
 
 *.ncl
 *.yaml
+*.txt
